Move `AnyTransactionReceipt` to `rpc-types-any`
