+++ crates/network/Cargo.toml
@@ -25,6 +25,7 @@ alloy-eips = { workspace = true, features = ["serde"] }
 alloy-json-rpc.workspace = true
 alloy-network-primitives.workspace = true
 alloy-primitives = { workspace = true, features = ["map"] }
+alloy-rpc-types-any = { workspace = true, features = ["serde"] }
 alloy-rpc-types-eth = { workspace = true, features = ["std", "serde"] }
 alloy-signer.workspace = true
 alloy-serde.workspace = true
