+++ crates/rpc-types-any/README.md
@@ -1,0 +1,3 @@
+# alloy-rpc-types-any
+
+Types for any network.
