+++ crates/rpc-types-any/src/transaction/mod.rs
@@ -1,0 +1,3 @@
+mod receipt;
+#[cfg(feature = "serde")]
+pub use receipt::AnyTransactionReceipt;
