+++ crates/rpc-types-eth/src/transaction/mod.rs
@@ -22,9 +22,6 @@ pub use error::ConversionError;
 mod receipt;
 pub use receipt::TransactionReceipt;
 
-#[cfg(feature = "serde")]
-pub use receipt::AnyTransactionReceipt;
-
 pub mod request;
 pub use request::{TransactionInput, TransactionRequest};
